# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Social Auto Funnel Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts and reels
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [post, reel, all]
            default: all
        - name: platform
          in: query
          schema:
            type: string
            enum: [facebook, instagram, all]
            default: all
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPosts"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List all automation rules
      parameters:
        - name: postId
          in: query
          schema:
            type: integer
        - name: isGlobal
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationRule"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create an automation rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleBody"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
  /rules/{id}:
    get:
      operationId: getRule
      tags: [rules]
      summary: Get a rule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
    put:
      operationId: updateRule
      tags: [rules]
      summary: Update an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRuleBody"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
    delete:
      operationId: deleteRule
      tags: [rules]
      summary: Delete an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /rules/{id}/toggle:
    post:
      operationId: toggleRule
      tags: [rules]
      summary: Enable or disable an automation rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleRuleBody"
      responses:
        "200":
          description: Rule toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRule"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        metaId:
          type: string
        platform:
          type: string
          enum: [facebook, instagram]
        type:
          type: string
          enum: [post, reel]
        caption:
          type: string
        thumbnailUrl:
          type: string
        permalink:
          type: string
        commentCount:
          type: integer
        likeCount:
          type: integer
        automationEnabled:
          type: boolean
        rulesCount:
          type: integer
        publishedAt:
          type: string
          format: date-time
        syncedAt:
          type: string
          format: date-time
      required: [id, tenantId, metaId, platform, type, commentCount, likeCount, automationEnabled, rulesCount, publishedAt, syncedAt]
    PaginatedPosts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    AutomationRule:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        postId:
          type: integer
          nullable: true
        isGlobal:
          type: boolean
        name:
          type: string
        keywords:
          type: array
          items:
            type: string
        replyText:
          type: string
        dmMessage:
          type: string
          nullable: true
        sendDm:
          type: boolean
        isEnabled:
          type: boolean
        matchCount:
          type: integer
        replyCount:
          type: integer
        dmCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tenantId, isGlobal, name, keywords, replyText, sendDm, isEnabled, matchCount, replyCount, dmCount, createdAt, updatedAt]
    CreateRuleBody:
      type: object
      properties:
        postId:
          type: integer
          nullable: true
        isGlobal:
          type: boolean
          default: false
        name:
          type: string
        keywords:
          type: array
          items:
            type: string
        replyText:
          type: string
        dmMessage:
          type: string
          nullable: true
        sendDm:
          type: boolean
          default: false
      required: [name, keywords, replyText]
    UpdateRuleBody:
      type: object
      properties:
        name:
          type: string
        keywords:
          type: array
          items:
            type: string
        replyText:
          type: string
        dmMessage:
          type: string
          nullable: true
        sendDm:
          type: boolean
        isEnabled:
          type: boolean
    ToggleRuleBody:
      type: object
      properties:
        isEnabled:
          type: boolean
      required: [isEnabled]
...